# doc/openapi/openapi-v1.yaml
openapi: 3.0.0
info:
  title: Data Series Tracker API
  version: 1.0.0
  description: |
    Public API provided by `Data Series Tracker`.
tags:
  - name: Datapoint
    description: Datapoint management.
  - name: Statistics
    description: Computed statistics.
paths:
  /datapoints:
    post:
      operationId: create-datapoint
      summary: 'Store a new datapoint.'
      tags:
        - Datapoint
      requestBody:
        $ref: '#/components/requestBodies/PostDatapointRequest'
      responses:
        '201':
          $ref: '#/components/responses/PostDatapointResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /devices/{device}/datapoints:
    parameters:
      - $ref: '#/components/parameters/device'
    delete:
      operationId: delete-device-datapoints
      summary: 'Delete all datapoints associated with a given device.'
      tags:
        - Datapoint
      responses:
        '204':
          description: 'Datapoints deleted successfully.'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /users/{user}/datapoints:
    parameters:
      - $ref: '#/components/parameters/user'
    delete:
      operationId: delete-user-datapoints
      summary: 'Delete all datapoints associated with a given user.'
      tags:
        - Datapoint
      responses:
        '204':
          description: 'Datapoints deleted successfully.'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /statistics/devices/{device}/avg:
    parameters:
      - $ref: '#/components/parameters/device'
    get:
      operationId: get-device-averages
      summary: 'Return list of 15 minute averages for a device.'
      tags:
        - Statistics
      responses:
        '200':
          $ref: '#/components/responses/GetAveragesResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /statistics/devices/{device}/moving_avg:
    parameters:
      - $ref: '#/components/parameters/device'
      - $ref: '#/components/parameters/windowSize'
    get:
      operationId: get-device-moving-averages
      summary: 'Return list of 15 minute moving averages for a device.'
      tags:
        - Statistics
      responses:
        '200':
          $ref: '#/components/responses/GetAveragesResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /statistics/users/{user}/avg:
    parameters:
      - $ref: '#/components/parameters/user'
    get:
      operationId: get-user-averages
      summary: 'Return list of 15 minute averages for a user.'
      tags:
        - Statistics
      responses:
        '200':
          $ref: '#/components/responses/GetAveragesResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /statistics/users/{user}/moving_avg:
    parameters:
      - $ref: '#/components/parameters/user'
      - $ref: '#/components/parameters/windowSize'
    get:
      operationId: get-user-moving-averages
      summary: 'Return list of 15 minute moving averages for a user.'
      tags:
        - Statistics
      responses:
        '200':
          $ref: '#/components/responses/GetAveragesResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
components:
  parameters:
    user:
      name: user
      description: 'User identifier.'
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/userId'
    device:
      name: device
      description: 'Device identifier.'
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/deviceId'
    windowSize:
      name: window_size
      description: 'Window size.'
      in: query
      required: true
      example: 2
      schema:
        type: integer
        format: int32
        minimum: 1
  schemas:
    userId:
      type: string
      example: 'harkonnen'
      minLength: 1
      maxLength: 50
      pattern: '^[a-zA-Z0-9]+$'
    deviceId:
      type: string
      example: 'SpaceballOne'
      minLength: 1
      maxLength: 50
      pattern: '^[a-zA-Z0-9]+$'
    timestamp:
      type: string
      format: 'date-time'
      example: '2020-08-15T10:15:27.000+02:00'
    value:
      type: number
      format: 'double'
      example: 20.04
    Error:
      type: object
      required:
        - error_code
        - error_description
      properties:
        error_code:
          type: string
          example: 'VALIDATION_ERROR'
        error_description:
          type: string
          example: 'Validation failed.'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorItem'
    ErrorItem:
      type: object
      required:
        - error_code
      properties:
        error_code:
          type: string
          example: 'FIELD_INVALID_VALUE'
        subject:
          type: string
          example: 'user'
        details:
          type: string
          example: 'The value of field "user" contains invalid character: "#".'
    Datapoint:
      type: object
      required:
        - timestamp
        - value
        - device
        - user
      properties:
        timestamp:
          $ref: '#/components/schemas/timestamp'
        value:
          $ref: '#/components/schemas/value'
        device:
          $ref: '#/components/schemas/deviceId'
        user:
          $ref: '#/components/schemas/userId'
    Average:
      type: object
      required:
        - timestamp
        - value
      properties:
        timestamp:
          $ref: '#/components/schemas/timestamp'
        value:
          $ref: '#/components/schemas/value'
  requestBodies:
    PostDatapointRequest:
      description: Request body related to operation `Post datapoint`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Datapoint'
  responses:
    ErrorResponse:
      description: Some error occurred. Please look at response body for more details.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PostDatapointResponse:
      description: Response body for operation `Post datapoint`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Datapoint'
    GetAveragesResponse:
      description: Response body containing computed average values.
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Average'